+++ goweb.go
@@ -1,57 +1,57 @@
 package main
 
 import (
-	"fmt"
-	"github.com/RobyFerro/go-web-framework/console"
-	console2 "github.com/RobyFerro/go-web/app/console"
-	"github.com/RobyFerro/go-web/app/kernel"
-	"github.com/common-nighthawk/go-figure"
+	gwf "github.com/RobyFerro/go-web-framework"
+	"github.com/RobyFerro/go-web/app"
+	"github.com/RobyFerro/go-web/app/console"
+	"github.com/RobyFerro/go-web/app/http/controller"
+	"github.com/RobyFerro/go-web/app/http/middleware"
+	"github.com/RobyFerro/go-web/database/model"
+	"github.com/RobyFerro/go-web/service"
 	"os"
-	"reflect"
+)
+
+var (
+	// Controllers will handle all Go-Web controller
+	Controllers = gwf.ControllerRegister{
+		List: []interface{}{
+			&controller.UserController{},
+			&controller.AuthController{},
+			&controller.HomeController{},
+		},
+		// Here is where you've to register your custom controller
+	}
+	// Models will handle all application models
+	Models = gwf.ModelRegister{
+		List: []interface{}{
+			model.User{},
+			model.FailedJob{},
+			// Here is where you've to register your custom models
+		},
+	}
+	// Services will handle all app services
+	Services = gwf.ServiceRegister{
+		List: []interface{}{
+			app.Configuration,
+			service.ConnectDB,
+			service.ConnectElastic,
+			service.ConnectMongo,
+			service.ConnectRedis,
+			// Here is where you've register your custom service
+		},
+	}
+	// Configuration represent all Go-Web application conf
+	Commands = gwf.CommandRegister{
+		List: map[string]interface{}{
+			"queue:failed": &console.QueueFailed{},
+			"queue:run":    &console.QueueRun{},
+			// Here is where you've to register your custom commands
+		},
+	}
 )
 
 // Main Go-Web entry point.
 // Service container will be passed as parameter for every method
 func main() {
-	var args string
-	pwd, _ := os.Getwd()
-	_ = os.Setenv("base_path", pwd)
-
-	// Start HTTP Kernel
-	httpKernel := kernel.StartKernel()
-
-	// Print Go-Web container
-	printHeader()
-	// Merge commands
-	mixCommands()
-
-	// New command handler
-	cmd := console.Commands[os.Args[1]]
-	if cmd == nil {
-		fmt.Println("Command not found!")
-		os.Exit(1)
-	}
-
-	v := reflect.ValueOf(cmd).MethodByName("Run")
-
-	if len(os.Args) == 3 {
-		args = os.Args[2]
-	}
-
-	v.Call([]reflect.Value{reflect.ValueOf(httpKernel), reflect.ValueOf(args), reflect.ValueOf(console.Commands)})
-}
-
-// Print Go-Web CLI header
-func printHeader() {
-	myFigure := figure.NewFigure("Go-Web", "graffiti", true)
-	myFigure.Print()
-
-	fmt.Println("Go-Web CLI tool - Author: roberto.ferro@ikdev.eu")
-}
-
-// Merge framework commands with yours
-func mixCommands() {
-	for i, c := range console2.Commands {
-		console.Commands[i] = c
-	}
+	gwf.Start(os.Args[1:], Commands, Controllers, Services, middleware.Middleware{}, Models)
 }