+++ service/db_connection.go
@@ -3,12 +3,13 @@ package service
 import (
 	"fmt"
 	"github.com/RobyFerro/go-web-framework"
+	"github.com/RobyFerro/go-web/app"
 	"github.com/jinzhu/gorm"
 	_ "github.com/jinzhu/gorm/dialects/mysql"
 )
 
 // Connect to sql database
-func ConnectDB(conf *gwf.Conf) *gorm.DB {
+func ConnectDB(conf *app.Conf) *gorm.DB {
 	connectionString, driver := createConnectionString(conf)
 	db, err := gorm.Open(driver, connectionString)
 
@@ -20,7 +21,7 @@ func ConnectDB(conf *gwf.Conf) *gorm.DB {
 }
 
 // Create string for SQL connection
-func createConnectionString(conf *gwf.Conf) (string, string) {
+func createConnectionString(conf *app.Conf) (string, string) {
 	return fmt.Sprintf(
 		"%s:%s@(%s:%d)/%s?charset=utf8&parseTime=True&loc=Local",
 		conf.Database.User,