+++ service/mongo.go
@@ -4,13 +4,14 @@ import (
 	"context"
 	"fmt"
 	"github.com/RobyFerro/go-web-framework"
+	"github.com/RobyFerro/go-web/app"
 	"go.mongodb.org/mongo-driver/mongo"
 	"go.mongodb.org/mongo-driver/mongo/options"
 	"time"
 )
 
 // Connect to MongoDB
-func ConnectMongo(conf *gwf.Conf) *mongo.Database {
+func ConnectMongo(conf *app.Conf) *mongo.Database {
 	client, err := mongo.NewClient(options.Client().ApplyURI(fmt.Sprintf("mongodb://%s:%d", conf.Mongo.Host, conf.Mongo.Port)))
 
 	if err != nil {