+++ service/redis.go
@@ -3,11 +3,12 @@ package service
 import (
 	"fmt"
 	"github.com/RobyFerro/go-web-framework"
+	"github.com/RobyFerro/go-web/app"
 	"github.com/go-redis/redis/v7"
 )
 
 // Connect to Redis
-func ConnectRedis(conf *gwf.Conf) *redis.Client {
+func ConnectRedis(conf *app.Conf) *redis.Client {
 	client := redis.NewClient(&redis.Options{
 		Addr:     fmt.Sprintf("%s:%d", conf.Redis.Host, conf.Redis.Port),
 		Password: conf.Redis.Password,